move unwatch into master only commands
Summary
WHAT!? you state. Well the truth is UNWATCH only makes
sense in a transaction state. But you can call UNWATCH when
not in transaction state. I do not believe when that happens
it should start a transaction. I also do not believe it should
end a transaction. So what to do with it then?